Add NodesListFilterMixin to handle root and parent filters.
- Add NodesListFilterMixin to node, user, and institution nodes list
views.
